Improve support of the zfs config file
Before this commit, the file at /etc/default/zfs was ignored.
If ZFS_MOUNT was not equal to one of the four true values
(yes, on, true, or 1) the mount command was still executed.
Same with share.
The code also precluded the possibility of not using the default
cache file.